# docs/superpowers/specs/messaging-api.openapi.yaml
openapi: 3.1.0
info:
  title: Hermes Messaging API
  version: 2.8.0
  description: |
    Private API bridging mobile clients to Hermes.

    **v2.8.0 changes:** photo attachments in chat. `POST /attachments` stages image
    uploads (JPEG, PNG, HEIC/HEIF; max 20 MB). `POST /conversations/{id}/messages`
    accepts optional `attachment_ids` (1–10) with optional caption. `Message.attachments[]`
    on list, sync, and create responses. `GET /attachments/{id}?variant=original|thumb|vision`
    serves JWT-scoped bytes. Caption-only `PATCH` on photo messages. Hermes receives
    multimodal history with vision JPEG derivatives. Requires companion app update.
    See `docs/superpowers/specs/2026-06-21-companion-photo-sharing-design.md`.

    **v2.7.0 changes:** structured tooling lines for companion process UX.
    `ToolingLine` replaces flat `kind: reasoning|tool` with Hermes-native
    `phase` (`reasoning` | `activity` | `status`), optional `tool`, and
    optional `args`. Session `tooling` events and persisted `message.process.lines`
    use the new shape. Interim assistant narration (e.g. before `memory` /
    `skill_manage`) appears as `phase: status` in the tooling lane. Icons are
    not sent — iOS maps `tool` + `args` locally. Requires companion app update.
    See `docs/superpowers/specs/2026-06-21-companion-tooling-lines-design.md`.

    **v2.6.0 changes:** per-device sync inbox. `PUT /devices/me` registers a stable
...

...

...
    `pct`, `g`, `ml`). No new routes or MCP tools. See
...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_credentials
            - invalid_token
            - weak_password
            - username_taken
            - not_found
            - run_conflict
            - no_active_run
            - edit_not_allowed
            - day_finalized
            - sync_marker_invalid
...
    ToolingLine:
      type: object
      required: [phase, text]
      properties:
        phase:
          type: string
          enum: [reasoning, activity, status]
          description: |
            reasoning — model thinking (committed line).
            activity — Hermes tool execution (`tool` + `args` when available).
            status — interim assistant narration before/during housekeeping;
            `tool` set when correlated with the next Hermes tool in the turn.
        text:
          type: string
          description: |
            Display copy. Prefer Hermes tool-progress `label` when the backend
            has it; not used as the type key for client grouping.
        tool:
          type: [string, 'null']
          description: |
            Hermes tool identifier when known (e.g. skill_view, memory,
            web_search, terminal, mcp_*). Null for reasoning-only lines or
            uncorrelated status lines.
        args:
          type: [object, 'null']
          additionalProperties: true
          description: |
            Optional subset of Hermes tool arguments for client presentation
            (e.g. memory action/target, skill_view name, web_search query).
            Omitted or null when unavailable.
    ProcessLine:
      description: |
        Alias for ToolingLine (persisted on assistant messages). v2.7.0+ uses
        `phase` / `tool` / `args`. v2.6 used `kind: reasoning|tool` only.
      allOf:
        - $ref: '#/components/schemas/ToolingLine'
    MessageProcess:
      type: object
      required: [lines]
      properties:
        lines:
          type: array
          items:
            $ref: '#/components/schemas/ProcessLine'
          description: Accumulated process trail for one assistant turn
    Message:
      type: object
      required: [id, conversation_id, role, content, created_at]
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        created_at:
          type: string
        process:
          $ref: '#/components/schemas/MessageProcess'
          description: Present on assistant messages when the turn had process activity
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/AttachmentSummary'
          description: Present on user messages with photos; omitted on text-only messages
    AttachmentSummary:
      type: object
      required: [id, content_type, byte_size, position]
      properties:
        id:
          type: string
          format: uuid
        content_type:
          type: string
          description: Original MIME type (e.g. image/heic)
        byte_size:
          type: integer
          minimum: 1
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        position:
          type: integer
          minimum: 0
        _links:
          type: object
          properties:
            self:
              $ref: '#/components/schemas/HalLink'
            thumb:
              $ref: '#/components/schemas/HalLink'
...
    CreateMessageRequest:
      type: object
      description: |
        Provide `text` (preferred) or `content` (legacy alias). At least one of
        non-empty `text`/`content` or `attachment_ids` is required. Optional
        `bootstrap` is accepted only on the **first message** of a conversation
        when no bootstrap is stored yet.
      properties:
        text:
          type: string
        content:
          type: string
        attachment_ids:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
          maxItems: 10
          description: Staged attachment ids from POST /attachments
        bootstrap:
          type: string
          minLength: 1
          maxLength: 4000
          description: |
            Hermes system bootstrap. Accepted only when this is the first
            message in the conversation and `bootstrap_prompt` is not already
            set. Ignored on subsequent messages. Not persisted as a message;
            never returned in GET responses.
    EditMessageRequest:
      type: object
      description: Provide `text` (preferred) or `content` (legacy alias).
      properties:
        text:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
    CreateMessageResponse:
      type: object
      required: [message]
      properties:
        message:
          $ref: '#/components/schemas/Message'
...
    HalLink:
      type: object
      required: [href]
      properties:
        href:
          type: string
          example: /conversations?limit=20&before=00000000-0000-4000-8000-000000000001
    HalLinks:
      type: object
      required: [self]
      properties:
        self:
          $ref: '#/components/schemas/HalLink'
        next:
          $ref: '#/components/schemas/HalLink'
        prev:
          $ref: '#/components/schemas/HalLink'
...
    MessageListResponse:
      type: object
      required: [messages, _links]
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        _links:
          $ref: '#/components/schemas/HalLinks'
...
    SseTokenEvent:
      type: object
      required: [text]
      properties:
        text:
          type: string
    SseProcessEvent:
      description: |
        Committed tooling line on the deprecated per-conversation stream.
        Same shape as ToolingLine (v2.7.0+).
      allOf:
        - $ref: '#/components/schemas/ToolingLine'
    SseProcessTokenEvent:
      type: object
      required: [kind, text]
      properties:
        kind:
          type: string
          const: reasoning
        text:
          type: string
          description: Reasoning delta; append to the in-flight reasoning line in the process block
    SseProcessCompleteEvent:
      type: object
      description: Process phase ended; reply tokens follow
      properties: {}
    SseDoneEvent:
      type: object
      required: [messageId]
      properties:
        messageId:
          type: string
          format: uuid
    SseErrorEvent:
      type: object
      required: [code]
      properties:
        code:
          type: string
          example: hermes_stream_failed
    SseTitleEvent:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 80
          description: Auto-generated title saved to the conversation
    SseRewindEvent:
      type: object
      required: [removedMessageIds]
      properties:
        removedMessageIds:
          type: array
          items:
            type: string
            format: uuid
          description: Message IDs removed from the transcript (e.g. superseded assistant reply)
...
paths:
...
  /conversations/{id}/messages:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [messages]
      summary: List messages
      description: |
        Returns a paginated message window. Default (no anchors) returns the
        **tail** — the last `limit` messages in chronological order within the
        page. Follow `_links.prev` for older history and `_links.next` toward
        the present.
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: before
          in: query
          description: Message id — return older messages before this anchor
          schema:
            type: string
            format: uuid
        - name: after
          in: query
          description: Message id — return newer messages after this anchor
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Paginated messages (chronological within page)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
        '400':
          description: Invalid pagination query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or not owned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [messages]
      summary: Send user message
      description: |
        Persists the user message, starts assistant run, returns 202 immediately.

        On the **first message** of a conversation with `title: null`, the
        backend also starts background title generation (parallel Hermes call).
        Connected stream clients may receive a `title` SSE event.

        Optional `bootstrap` on the first message stores Hermes skill-routing
        instructions for the conversation (write-only; not returned in GET).
        Ignored on second and later messages. Invalid `bootstrap` (empty or
        longer than 4000 characters) returns `400 invalid_request`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
      responses:
        '202':
          description: Message accepted, run started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateMessageResponse'
        '400':
          description: Invalid or empty message, or invalid bootstrap on first message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or not owned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Active run already in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /conversations/{id}/messages/{messageId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: messageId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    patch:
      tags: [messages]
      summary: Edit latest user message and re-run assistant
      description: |
        Edits a user message **in place** when it is the second-to-last message in
        the transcript and the last message is a completed assistant reply. No active
        run may be in progress.

        Atomically:
        - deletes the superseded assistant message and its run record
        - updates the user message content (same message id)
        - rotates `hermes_session_id` to discard Hermes session context
        - starts a new assistant run

        Returns `202` immediately. Client opens `GET /conversations/{id}/stream`
        to receive `rewind`, then `token` / `done` as usual.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditMessageRequest'
      responses:
        '202':
          description: Message updated, replacement run started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateMessageResponse'
        '400':
          description: Invalid text or message not in editable position
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_request:
                  value: { error: invalid_request }
                edit_not_allowed:
                  value: { error: edit_not_allowed }
        '404':
          description: Not found or not owned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Active run already in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /conversations/{id}/stream:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [messages]
      deprecated: true
      summary: SSE stream for active assistant run
      description: |
        **Deprecated:** use `GET /events/stream` instead. Per-conversation streams
        attach after send and miss early tooling events; the session stream
        eliminates that race.

        Server-Sent Events for the current active run. Attaches only to the
        in-flight run; no replay. Stream ends on `done` or `error`.

        **Two-phase assistant stream:** `process_token` streams reasoning text
        incrementally; `process` emits committed `ToolingLine` rows (`phase`,
        `text`, optional `tool` / `args`); `process_complete` ends the process
        block; `token` events carry the final reply. Instant replies may omit
        the process phase.

        **SSE events:**

        | event | data schema | when |
        |-------|-------------|------|
        | `rewind` | `SseRewindEvent` | Messages removed before a replacement run (message edit) |
        | `process_token` | `SseProcessTokenEvent` | Reasoning delta while Hermes is thinking |
        | `process` | `SseProcessEvent` | Committed tooling line (v2.7.0 ToolingLine shape) |
        | `process_complete` | `SseProcessCompleteEvent` | Process phase ended; collapse process UI before reply |
        | `token` | `SseTokenEvent` | Assistant reply text chunk |
        | `title` | `SseTitleEvent` | Auto-generated title saved (first message only) |
        | `done` | `SseDoneEvent` | Assistant message persisted |
        | `error` | `SseErrorEvent` | Run failed |
      security:
        - bearerAuth: []
      x-sse-events:
        rewind:
          $ref: '#/components/schemas/SseRewindEvent'
        process_token:
          $ref: '#/components/schemas/SseProcessTokenEvent'
        process:
          $ref: '#/components/schemas/SseProcessEvent'
        process_complete:
          $ref: '#/components/schemas/SseProcessCompleteEvent'
        token:
          $ref: '#/components/schemas/SseTokenEvent'
        title:
          $ref: '#/components/schemas/SseTitleEvent'
        done:
          $ref: '#/components/schemas/SseDoneEvent'
        error:
          $ref: '#/components/schemas/SseErrorEvent'
      responses:
        '200':
          description: Event stream
          content:
            text/event-stream:
              schema:
                type: string
        '404':
          description: Not found or not owned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: No active run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...